ORDER-0: add security layer (#2)
* ORDER-0: change defaault port

* ORDER-0: add security layer

* ORDER-0: IT fix

* ORDER-0: implement getOrder(ById) fix + polishing

* ORDER-0: implement IT fix

+++ src/main/java/com/innowise/orderservice/controller/ItemController.java
@@ -2,7 +2,9 @@ package com.innowise.orderservice.controller;
 
 import com.innowise.orderservice.model.dto.item.ItemRequestDto;
 import com.innowise.orderservice.model.dto.item.ItemResponseDto;
+import com.innowise.orderservice.service.AccessPolicyService;
 import com.innowise.orderservice.service.ItemService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -26,36 +28,47 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ItemController {
     private final ItemService itemService;
+    private final AccessPolicyService accessPolicyService;
 
     @PostMapping
-    public ResponseEntity<ItemResponseDto> createItem(@Valid @RequestBody ItemRequestDto request) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(itemService.createItem(request));
+    public ResponseEntity<ItemResponseDto> createItem(
+            HttpServletRequest request,
+            @Valid @RequestBody ItemRequestDto itemRequest
+    ) {
+        accessPolicyService.requireUserOrAdmin(request);
+        return ResponseEntity.status(HttpStatus.CREATED).body(itemService.createItem(itemRequest));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ItemResponseDto> getItemById(@PathVariable Long id) {
+    public ResponseEntity<ItemResponseDto> getItemById(HttpServletRequest request, @PathVariable Long id) {
+        accessPolicyService.requireUserOrAdmin(request);
         return ResponseEntity.ok(itemService.getItemById(id));
     }
 
     @GetMapping
     public ResponseEntity<Page<ItemResponseDto>> getItems(
+            HttpServletRequest request,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size
     ) {
+        accessPolicyService.requireUserOrAdmin(request);
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "createdAt"));
         return ResponseEntity.ok(itemService.getItems(pageable));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<ItemResponseDto> updateItemById(
+            HttpServletRequest request,
             @PathVariable Long id,
-            @Valid @RequestBody ItemRequestDto request
+            @Valid @RequestBody ItemRequestDto itemRequest
     ) {
-        return ResponseEntity.ok(itemService.updateItemById(id, request));
+        accessPolicyService.requireAdmin(request);
+        return ResponseEntity.ok(itemService.updateItemById(id, itemRequest));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteItemById(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteItemById(HttpServletRequest request, @PathVariable Long id) {
+        accessPolicyService.requireAdmin(request);
         itemService.deleteItemById(id);
         return ResponseEntity.noContent().build();
     }
